Move extensions to sample.
