+++ sample/src/main/kotlin/com/uber/autodispose/sample/KotlinActivity.kt
@@ -22,6 +22,7 @@ import android.support.v7.app.AppCompatActivity
 import android.util.Log
 import com.uber.autodispose.android.lifecycle.AndroidLifecycleScopeProvider
 import com.uber.autodispose.kotlin.autoDisposeWith
+import com.uber.autodispose.recipes.subscribeBy
 import io.reactivex.Observable
 import java.util.concurrent.TimeUnit
 
@@ -45,7 +46,7 @@ class KotlinActivity : AppCompatActivity() {
     Observable.interval(1, TimeUnit.SECONDS)
         .doOnDispose { Log.i(TAG, "Disposing subscription from onCreate()") }
         .autoDisposeWith(scopeProvider)
-        .subscribe { num -> Log.i(TAG, "Started in onCreate(), running until onDestroy(): " + num) }
+        .subscribeBy { Log.i(TAG, "Started in onCreate(), running until onDestroy(): $it") }
   }
 
   override fun onStart() {
@@ -58,7 +59,7 @@ class KotlinActivity : AppCompatActivity() {
     Observable.interval(1, TimeUnit.SECONDS)
         .doOnDispose { Log.i(TAG, "Disposing subscription from onStart()") }
         .autoDisposeWith(scopeProvider)
-        .subscribe { num -> Log.i(TAG, "Started in onStart(), running until in onStop(): " + num) }
+        .subscribeBy { Log.i(TAG, "Started in onStart(), running until in onStop(): $it") }
   }
 
   override fun onResume() {
@@ -71,9 +72,7 @@ class KotlinActivity : AppCompatActivity() {
     Observable.interval(1, TimeUnit.SECONDS)
         .doOnDispose { Log.i(TAG, "Disposing subscription from onResume()") }
         .autoDisposeWith(scopeProvider)
-        .subscribe { num ->
-          Log.i(TAG, "Started in onResume(), running until in onPause(): " + num)
-        }
+        .subscribeBy { Log.i(TAG, "Started in onResume(), running until in onPause(): $it") }
 
     // Setting a specific untilEvent, this should dispose in onDestroy.
     Observable.interval(1, TimeUnit.SECONDS)
@@ -81,9 +80,7 @@ class KotlinActivity : AppCompatActivity() {
           Log.i(TAG, "Disposing subscription from onResume() with untilEvent ON_DESTROY")
         }
         .autoDisposeWith(AndroidLifecycleScopeProvider.from(this, Lifecycle.Event.ON_DESTROY))
-        .subscribe{ num ->
-          Log.i(TAG, "Started in onResume(), running until in onDestroy(): " + num)
-        }
+        .subscribeBy { Log.i(TAG, "Started in onResume(), running until in onDestroy(): $it") }
   }
 
   override fun onPause() {
